Update cookieDomainRewrite setting to allow top level domain to remain
The setting of "" removes the domain value, which breaks domain cookies like `.ilr.cornell.edu` (used for the UTM code storage cookie).

But it's clear that the setting was added to allow for D7 session cookies, so the fallback (*) is included to remove all other cookie domains and allow for that.